server/requirements.txt
Flask
flask-cors
flask-sqlalchemy
PyJWT
Werkzeug
python-dotenv
psycopg2-binary
